chore(ci): add stencil v4-next workflow (#27598)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

we have no way to test stencil v4 as a part of the nightly build

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

this commit adds a new ci cron job, 'stencil v4 build'. the purpose of
this job is to run the ionic framework ci suite on an nightly basis,
similar to the stencil nightly build.

a separate workflow has been created, as opposed to adding a `matrix` to
the pre-existing workflow, as doing so would require a great deal of
refactoring to the existing workflow. instead, a separate workflow
allows us to simply delete the this workflow once stencil v4 is
released.

as a part of this commit, we add a new input to the
'build-core-stencil-prerelease' action, allowing users of the action to
input the npm tag of `@stencil/core` to install

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

I tested the Stencil nightly job manually -
https://github.com/ionic-team/ionic-framework/actions/runs/5180201562

I was unable to run the v4 job, as it's not in `main` (yet 😉 ). But, I
can diff it against the known-to-be-good nightly workflow. The following
shows only the name of the job & npm tag are different:
```diff
 diff stencil-v4.yml stencil-nightly.yml
3c3
< name: 'Stencil v4 Build'
---
> name: 'Stencil Nightly Build'
21c21
<   build-core-with-stencil-v4:
---
>   build-core-with-stencil-nightly:
27c27
<           stencil-version: v4-next
---
>           stencil-version: nightly
30c30
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
37c37
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
44c44
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
65c65
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
93c93
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
129c129
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
165c165
<     needs: [build-core-with-stencil-v4]
---
>     needs: [build-core-with-stencil-nightly]
```

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->