fix(toggle): add top and bottom margins when in ion-item (#27784)
Issue number: Part of #27498

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

When `ion-toggle` is used within `ion-item`, top and bottom margins are
not added to the label, while they are present when using the legacy
syntax. We didn't catch this because the issue is only visible when
using a label that breaks onto more than one line; otherwise, the height
of the item exceeds what the margins would've added.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

Margins added. Values were taken from `ion-item` styles
[here](https://github.com/ionic-team/ionic-framework/blob/096d9cc931dfe3a1b4b39d6652112ef209be7147/core/src/components/item/item.ios.scss#L203-L206)
and
[here](https://github.com/ionic-team/ionic-framework/blob/096d9cc931dfe3a1b4b39d6652112ef209be7147/core/src/components/item/item.md.scss#L298-L301),
which no longer get applied because the `ion-label` is nested in an
additional shadow component. Note that left/right margins are already
included in the modern syntax labels.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

I noticed that this test is incorrectly named `label.e2e.ts` instead of
`toggle.e2e.ts` while working on this. I'll open a separate PR to
correct that once these changes are merged into `main`, to avoid merge
conflicts.

---------

Co-authored-by: ionitron <hi@ionicframework.com>